+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 CERTitude - Changelog
 =============
 
+* 2017-03-15: New visualization based on Bokeh for large results
 * 2017-03-15: Added the "select" feature which allows retrieving metadata about the items discovered on the target
 * 2017-03-15: CHANGELOG creation